Use bound python generics to represent NotFound
Instead of passing a class around as a prop, we can represent it
as a generic! And by binding the generic type to only be repos,
we are now able to represent that what we have not found is always
a repo of some sort purely within the type system!

Also we no longer have 'NotFound' be itself a type of 'Repo',
because it is not. It is a wrapper / container for a particular class
of repo that was not found.