Add a simple 'Guess' resolver that supports git repos
Sometimes we just don't know what a URL is, and have to really
guess. The Guess resolver does this, and returns other Repos that
can then be recursively resolved.

Will need to figure out how it can support other things (like dataverse)
to inform how the design progresses. But this seems like a good start.

Note that this doesn't allow for parsing things like branch names
from the URL, as we don't know - so `Git` is just assumed to be `HEAD`.